+++ pkg/decoder/smartlabel/v1/decoder_test.go
@@ -64,7 +64,7 @@ func TestDecode(t *testing.T) {
 
 	tests := []struct {
 		payload     string
-		port        int16
+		port        uint8
 		devEui      string
 		expected    any
 		expectedErr string
@@ -272,7 +272,7 @@ func TestPayloadTooLong(t *testing.T) {
 func TestFeatures(t *testing.T) {
 	tests := []struct {
 		payload string
-		port    int16
+		port    uint8
 	}{
 		{
 			payload: "0f501079",
