+++ pkg/decoder/tagsl/v1/decoder_test.go
@@ -15,7 +15,7 @@ import (
 func TestDecode(t *testing.T) {
 	tests := []struct {
 		payload        string
-		port           int16
+		port           uint8
 		autoPadding    bool
 		skipValidation bool
 		expected       any
@@ -1269,7 +1269,7 @@ func TestDecode(t *testing.T) {
 func TestValidationErrors(t *testing.T) {
 	tests := []struct {
 		payload  string
-		port     int16
+		port     uint8
 		expected error
 	}{
 		{
@@ -1612,7 +1612,7 @@ func TestFullDecode(t *testing.T) {
 		skipValidation bool
 		expectedData   any
 		expectedStatus *Status
-		port           int16
+		port           uint8
 	}{
 		{
 			payload: "8002cdcd1300744f5e166018040b14341a",
@@ -1782,7 +1782,7 @@ func TestPayloadTooLong(t *testing.T) {
 func TestFeatures(t *testing.T) {
 	tests := []struct {
 		payload        string
-		port           int16
+		port           uint8
 		skipValidation bool
 	}{
 		{
@@ -1977,7 +1977,7 @@ func TestFeatures(t *testing.T) {
 func TestMarshal(t *testing.T) {
 	tests := []struct {
 		payload  string
-		port     int16
+		port     uint8
 		expected []string
 	}{
 		{
